Ensure NVMe journal partition is detected after labeling
After labeling the journal partition, call partprobe and wait for udev
to settle to ensure that the udev rules create the proper /dev/disk
symbolic links.

This should be done to ensure that the followup conditional wipe
operation has the proper symbolic link in place so it doesn't
inadvertently attempt to wipe the whole disk.

Closes-Bug: #1913343
Signed-off-by: Robert Church <robert.church@windriver.com>